refactor FastFormalaItem parts
