+++ agave/src/Interop/RangeCaches.ts
@@ -5,11 +5,12 @@
 // this cache can be setup when we load (and are refreshed), and referenced to get the ranges
 import { BracketDefBuilder } from "../Brackets/BracketDefBuilder";
 import { GameDataSources } from "../Brackets/GameDataSources";
-import { FastFormulaAreasItems, FastFormulaAreas } from "./FastFormulaAreas/FastFormulaAreas";
+import { FastFormulaAreas } from "./FastFormulaAreas/FastFormulaAreas";
 import { JsCtx } from "./JsCtx";
 import { RangeInfo, Ranges } from "./Ranges";
 import { RulesBuilder } from "../Brackets/RulesBuilder";
 import { s_staticConfig } from "../StaticConfig";
+import { FastFormulaAreasItems } from "./FastFormulaAreas/FastFormulaAreasItems";
 
 export class RangeCacheItemType
 {
